fixing problems and adding patterns
Trying to add repository pattern to create loose coupling, and further separating the concern with data handling, by using DI. Also added generics for reuseability for later.  Also fixed some issues with the splitting of projects in solutions, and fixed the name so that it will be consistent throughout. One thing though is that migration got to have a --project and --startup to work