Added more repo, changed to areabased, added product section, new db entry
Added more repo, changed the layout of the project to be based on area, and each area have their own controllers. This is to create different sections for admins and customer(TODO: also create another for seller). I have also created new MVC sections for both of them, as well as creating something called Product, and also created model to migrate it to a database.