foreign keys, new migration and added dropdown menu for categories in product create page
I have added a foreign key restriction inside the product model, so that I can refer to what category it belongs to. I have created a new migration for it, as a well as an imagelink value. I have also change the front end to include a list of all Categories(as a dropdown) when the user adds a new product.