Created details page, added ASP.Net core Identity, added alert when deleting product, recreated delete and getall products
I have chosen to create an API for the delete and get all because alert thirdparty library integration and easier to load every product(Returning Json data). Then I also created a landing page in the homecontroller, where all the product is displayed in a list, with details and also a details page that will display more details if you click on it. Lastly it took a bit of time, but I successfully integrated asp.net core identity into my project, and also changed the db context, and created migration adding the nessesary tables in the db, as well as a custom IdentityUser. Note you there are quite a few tables that will be added to the database. Lastly I just added some changes to the layout for a more consistent feel.