Created roles for accounts and areas restriction
Created model for different roles and was able to put into the database, if it detects that no row in the role table. Also added the ability for giving a role to a user when they register, and right now, only those with admin rights have access to the Category and Product page. Added a new dropdown(select box) in the register page. Fixed bug with not being able to land on the right page when you register, fixed the error that didn't let you register because of the email token.