Added stripe, order confirmation page, ordering system, ordering system inside database,
I have successfully crated a ordering system for the shopping cart. Where I have implemented Stripe and a system that can handle the stripe session to database, this is for status updates mainly. To accomplish that, I have added orderheader and orderdetail, where the orderheader is the information about customer and status, where orderdetails is details about each product that is in the order. Then also created a cartcontroller to handle the cart pages as well as summary and order.