Added controller for handling orders, as well as new api calls for getting orderlist
I have added added another controller and route on the platform that will handle and show all the orders that has happened and what status they are in, as well as added toggling of buttons, to show at appropriate order status for each order. If you are company that have made an order but haven't paid you will get a "pay now" button, and order that are in process will have a ship button together with provided carrier and trackingnumber. You also get a button for canceling orders. Created a details page where you will be redirected to if payment with the stripe api went through. Also fixed a bug where there was a logical error inside OrderHeaderRpository.cs that made it so that paymentintentid were never saved.  On the frontend side I used ajax to get all the orders, also I modified so that it can be grouped base on status, if the user doesn't want to look all the orders all the time.